Second part of file and directory rename for STM32
