--- src/HAL/STM32F4/STM32F4.ino
@@ -1,5 +1,0 @@
-// Placeholder file
-// Nothing to see here ...
-//
-// This file is only present so the Arduino IDE can edit the .h file(s)
-
